+++ dnn.h
@@ -51,6 +51,9 @@ int Layer_OutputNameToIndex(Layer layer, const char* name);
 const char* Layer_GetName(Layer layer);
 const char* Layer_GetType(Layer layer);
 
+void NMSBoxes(struct Rects bboxes, FloatVector scores, float score_threshold, float nms_threshold, IntVector* indices);
+void NMSBoxesWithParams(struct Rects bboxes, FloatVector scores, const float score_threshold, const float nms_threshold, IntVector* indices, const float eta, const int top_k);
+
 #ifdef __cplusplus
 }
 #endif
